Use original classes for Country deserializing

+++ src/main/java/com/maxmind/geoip2/model/EnterpriseResponse.java
@@ -108,7 +108,7 @@ public final class EnterpriseResponse extends AbstractCityResponse {
                     model.getTraits().getStaticIpScore()
                 ) :
                 new Traits(
-                    null,
+                    (Integer) null,
                     null,
                     null,
                     null,
